refactor: extract common functionality of param aware inputs to base class
Creates `AbstractParamAwareInput`, which:

- Defines the methods that have common signatures across symfony/console versions
- Adds a `modifyQuestion` abstract method.
- Extracts `getParam()`, and adds a call to `modifyQuestion()`
  immediately after pulling the question from the input parameter.

From here, I updated each of `TypeHintedParamAwareInput` and
`NonHintedParamAwareInput` to extend the new base class, remove the
common methods, and define `modifyQuestion()`. For
`NonHintedParamAwareInput`, `modifyQuestion()` is a no-op; for
`TypeHintedParamAwareInput`, it sets the max attempts on the question
instance provided to it.
